# Optimize `Base.callbacks_needed` by reordering checks